pyproject.toml
[project]
name = "strands-agents-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.34",
    "nest-asyncio>=1.6.0",
    "strands-agents>=1.9.1",
    "streamlit>=1.49.1",
]
